move module and add documentation
