+++ src/deepdisc/preprocessing/process.py
@@ -174,54 +174,4 @@ def write_scarlet_results(
 
 
 
-def fitsim_to_numpy(img_files, outdir, dset=None):
-        
-    for images in img_files:
-        full_im = []
-        for img in images:
-            with fits.open(img, memmap=False, lazy_load_hdus=False) as hdul:
-                data = hdul[0].data
-                full_im.append(data)
-
-        full_im = np.array(full_im)
-
-        tract = int(img.split('_')[1])
-        patch = (int(img.split('_')[2].split('_')[0][0]),int(img.split('_')[2].split('_')[0][-1]))
-        sp = int(img.split('_')[3])
-
-        np.save(os.path.join(outdir,f'{tract}_{patch[0]},{patch[1]}_{sp}_images.npy'),full_im)
-   
-    return 
-
-
-def fitsim_to_hdf5(img_files, outdir, dset='train'):
-    all_images=[]
-    for images in img_files:
-        full_im = []
-        for img in images:
-            with fits.open(img, memmap=False, lazy_load_hdus=False) as hdul:
-                data = hdul[0].data
-                full_im.append(data)
-
-        full_im = np.array(full_im)
-        all_images.append(full_im.flatten())
-    all_images = np.array(all_images)
-    
-    with h5py.File(os.path.join(outdir,f"flattened_images_{dset}.hdf5"), "w") as f:
-        data = f.create_dataset("images", data=all_images)
-   
-    return 
-
-
-
-def ddict_to_hdf5(dataset_dicts,outname):
-    
-    data = [json.dumps(this_dict) for this_dict in dataset_dicts]
-
-    dt = h5py.special_dtype(vlen=str)
-
-    dataset = file.create_dataset(outname, data=data, dtype=dt)
-    return
-
-
